requirements.txt
pyusb
hidapi
libusb